+++ packages/xchain-bitcoin/src/types/client-types.ts
@@ -1,9 +1,4 @@
-import { Address, FeeOptionKey, Fees, Network } from '@xchainjs/xchain-client'
-
-export type FeeRate = number
-export type FeeRates = Record<FeeOptionKey, FeeRate>
-
-export type FeesWithRates = { rates: FeeRates; fees: Fees }
+import { Address, Network } from '@xchainjs/xchain-client'
 
 export type NormalTxParams = { addressTo: string; amount: number; feeRate: number }
 export type VaultTxParams = NormalTxParams & { memo: string }