+++ packages/xchain-ethereum/CHANGELOG.md
@@ -12,6 +12,12 @@
 
 - Optional `gasLimitFallback` parameter for `approve` call
 
+
+# v.0.21.5 (2021-06-29)
+
+- added support for pulling fees from thornode.
+
+
 # v.0.21.4 (2021-06-07)
 
 ### Fix